Developer同步杨显运、刘杰、杜俊杰的合并进度

See merge request audiogroup/a6902/a6902_tx_rx!35

+++ applications/HL_A6902/module/display/page/hl_mod_page_monitor_set.c
@@ -53,7 +53,7 @@ LV_IMG_DECLARE(Other_monitor_tx_white);//
 
 static void hl_monitor_test_cb(hl_b_two_in_one_check_t event_num)
 {
-    uint8_t value = 0;
+    uint32_t value = 0;
     switch(event_num){
         case HL_B_TWO_ONE_CHECK_LEFT:
             value = MONITOR_UAC_IN;
