+++ applications/HL_A6902/module/display/page/hl_mod_page_sound_effect.c
@@ -52,7 +52,7 @@ LV_IMG_DECLARE(Other_sound_white);//声音增强
 
 static void hl_soundeffect_test_cb(hl_b_two_in_one_check_t event_num)
 {
-    uint8_t value = 0;
+    uint32_t value = 0;
     switch(event_num){
         case HL_B_TWO_ONE_CHECK_LEFT:
             value = HIGH_FIDELITY;
